+++ src/main/java/com/backend/allreva/AllrevaApplication.java
@@ -1,0 +1,13 @@
+package com.backend.allreva;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AllrevaApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AllrevaApplication.class, args);
+    }
+
+}
