+++ src/main/java/com/backend/allreva/common/config/SwaggerConfig.java
@@ -1,0 +1,16 @@
+package com.backend.allreva.common.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    public OpenAPI openAPI() {
+        return new OpenAPI()
+                .info(new Info().title("AllReva"));
+    }
+}
