+++ src/main/java/com/backend/allreva/common/exception/code/ErrorCode.java
@@ -1,0 +1,7 @@
+package com.backend.allreva.common.exception.code;
+
+public record ErrorCode(Integer status, String code, String message) {
+    public static ErrorCode of(final Integer status, final String code, final String message) {
+        return new ErrorCode(status, code, message);
+    }
+}
