+++ src/main/java/com/backend/allreva/common/exception/code/ErrorCodeInterface.java
@@ -1,0 +1,5 @@
+package com.backend.allreva.common.exception.code;
+
+public interface ErrorCodeInterface {
+    ErrorCode getErrorCode();
+}
