+++ src/main/java/com/backend/allreva/common/model/Email.java
@@ -1,0 +1,10 @@
+package com.backend.allreva.common.model;
+
+import jakarta.persistence.Embeddable;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Email {
+}
