+++ src/main/java/com/backend/allreva/member/command/application/ImageService.java
@@ -1,0 +1,4 @@
+package com.backend.allreva.member.command.application;
+
+public interface ImageService {
+}
