+++ src/main/java/com/backend/allreva/member/command/application/MemberCommandService.java
@@ -1,0 +1,9 @@
+package com.backend.allreva.member.command.application;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberCommandService {
+}
