+++ src/main/java/com/backend/allreva/member/command/application/dto/AddMemberRequest.java
@@ -1,0 +1,4 @@
+package com.backend.allreva.member.command.application.dto;
+
+public class AddMemberRequest {
+}
