+++ src/main/java/com/backend/allreva/member/command/domain/Member.java
@@ -1,0 +1,18 @@
+package com.backend.allreva.member.command.domain;
+
+import com.backend.allreva.common.model.Email;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Member {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Embedded
+    private Email email;
+}
