+++ src/main/java/com/backend/allreva/member/exception/MemberErrorCode.java
@@ -1,0 +1,21 @@
+package com.backend.allreva.member.exception;
+
+import com.backend.allreva.common.exception.code.ErrorCode;
+import com.backend.allreva.common.exception.code.ErrorCodeInterface;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@RequiredArgsConstructor
+public enum MemberErrorCode implements ErrorCodeInterface {
+    MEMBER_NOT_FOUND(HttpStatus.NOT_FOUND.value(), "MEMBER_NOT_FOUND", "존재하지 않는 회원입니다." )
+    ;
+
+    private final Integer status;
+    private final String errorCode;
+    private final String message;
+
+    @Override
+    public ErrorCode getErrorCode() {
+        return ErrorCode.of(status, errorCode, message);
+    }
+}
