+++ src/main/java/com/backend/allreva/member/exception/MemberNotFoundException.java
@@ -1,0 +1,9 @@
+package com.backend.allreva.member.exception;
+
+import com.backend.allreva.common.exception.CustomException;
+
+public class MemberNotFoundException extends CustomException {
+    public MemberNotFoundException() {
+        super(MemberErrorCode.MEMBER_NOT_FOUND);
+    }
+}
