+++ src/main/java/com/backend/allreva/member/infra/MemberRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.backend.allreva.member.infra;
+
+public interface MemberRepositoryCustom {
+}
