+++ src/main/java/com/backend/allreva/member/infra/MemberRepositoryImpl.java
@@ -1,0 +1,12 @@
+package com.backend.allreva.member.infra;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class MemberRepositoryImpl implements MemberRepositoryCustom{
+
+    private final JPAQueryFactory queryFactory;
+}
