+++ src/main/java/com/backend/allreva/member/query/MemberDetail.java
@@ -1,0 +1,4 @@
+package com.backend.allreva.member.query;
+
+public class MemberDetail {
+}
