+++ src/main/java/com/backend/allreva/member/query/MemberDetailRepository.java
@@ -1,0 +1,7 @@
+package com.backend.allreva.member.query;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberDetailRepository {
+}
