+++ src/main/java/com/backend/allreva/member/query/application/MemberQueryService.java
@@ -1,0 +1,9 @@
+package com.backend.allreva.member.query.application;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberQueryService {
+}
