+++ src/main/java/com/backend/allreva/member/ui/MemberController.java
@@ -1,0 +1,9 @@
+package com.backend.allreva.member.ui;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("api/v1/members")
+public class MemberController {
+}
