--- src/main/java/com/kwanse/allreva/AllrevaApplication.java
@@ -1,13 +1,0 @@
-package com.kwanse.allreva;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AllrevaApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AllrevaApplication.class, args);
-    }
-
-}
