--- src/main/java/com/kwanse/allreva/common/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package com.kwanse.allreva.common.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
