--- src/main/java/com/kwanse/allreva/common/exception/CustomException.java
@@ -1,16 +1,0 @@
-package com.kwanse.allreva.common.exception;
-
-import com.kwanse.allreva.common.exception.code.ErrorCode;
-import com.kwanse.allreva.common.exception.code.ErrorCodeInterface;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class CustomException extends RuntimeException {
-    private final ErrorCodeInterface errorCode;
-
-    public ErrorCode getErrorCode() {
-        return this.errorCode.getErrorCode();
-    }
-}
