--- src/main/java/com/kwanse/allreva/common/exception/code/ErrorCodeInterface.java
@@ -1,5 +1,0 @@
-package com.kwanse.allreva.common.exception.code;
-
-public interface ErrorCodeInterface {
-    ErrorCode getErrorCode();
-}
