--- src/main/java/com/kwanse/allreva/common/model/Email.java
@@ -1,10 +1,0 @@
-package com.kwanse.allreva.common.model;
-
-import jakarta.persistence.Embeddable;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Email {
-}
