--- src/main/java/com/kwanse/allreva/member/command/application/ImageService.java
@@ -1,4 +1,0 @@
-package com.kwanse.allreva.member.command.application;
-
-public interface ImageService {
-}
