--- src/main/java/com/kwanse/allreva/member/command/application/MemberCommandService.java
@@ -1,9 +1,0 @@
-package com.kwanse.allreva.member.command.application;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class MemberCommandService {
-}
