--- src/main/java/com/kwanse/allreva/member/command/application/dto/AddMemberRequest.java
@@ -1,4 +1,0 @@
-package com.kwanse.allreva.member.command.application.dto;
-
-public class AddMemberRequest {
-}
