--- src/main/java/com/kwanse/allreva/member/command/domain/Member.java
@@ -1,18 +1,0 @@
-package com.kwanse.allreva.member.command.domain;
-
-import com.kwanse.allreva.common.model.Email;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Embedded
-    private Email email;
-}
