--- src/main/java/com/kwanse/allreva/member/exception/MemberNotFoundException.java
@@ -1,9 +1,0 @@
-package com.kwanse.allreva.member.exception;
-
-import com.kwanse.allreva.common.exception.CustomException;
-
-public class MemberNotFoundException extends CustomException {
-    public MemberNotFoundException() {
-        super(MemberErrorCode.MEMBER_NOT_FOUND);
-    }
-}
