--- src/main/java/com/kwanse/allreva/member/infra/MemberRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.kwanse.allreva.member.infra;
-
-public interface MemberRepositoryCustom {
-}
