--- src/main/java/com/kwanse/allreva/member/infra/MemberRepositoryImpl.java
@@ -1,12 +1,0 @@
-package com.kwanse.allreva.member.infra;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class MemberRepositoryImpl implements MemberRepositoryCustom{
-
-    private final JPAQueryFactory queryFactory;
-}
