--- src/main/java/com/kwanse/allreva/member/query/MemberDetail.java
@@ -1,4 +1,0 @@
-package com.kwanse.allreva.member.query;
-
-public class MemberDetail {
-}
