--- src/main/java/com/kwanse/allreva/member/query/MemberDetailRepository.java
@@ -1,7 +1,0 @@
-package com.kwanse.allreva.member.query;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberDetailRepository {
-}
