--- src/main/java/com/kwanse/allreva/member/query/application/MemberQueryService.java
@@ -1,9 +1,0 @@
-package com.kwanse.allreva.member.query.application;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class MemberQueryService {
-}
