--- src/main/java/com/kwanse/allreva/member/ui/MemberController.java
@@ -1,9 +1,0 @@
-package com.kwanse.allreva.member.ui;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("api/v1/members")
-public class MemberController {
-}
