+++ src/test/java/com/backend/allreva/AllrevaApplicationTests.java
@@ -1,0 +1,13 @@
+package com.backend.allreva;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AllrevaApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
