--- src/test/java/com/kwanse/allreva/AllrevaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kwanse.allreva;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AllrevaApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
